Avoid passing FILE* across DLL boundary
TestApp was opening a file and then passing the FILE* to SpatialUnderstanding. According to https://msdn.microsoft.com/en-us/library/ms235460.aspx, this can result in errors because the DLL may have differing CRT state. This *seems* like it explains the issue that I was seeing (#88).

To fix this, I am passing the desired filename to a new exported SpatialUnderstanding function that runs both the Init and the Update portions of the test. And I am removing export of the previous Init and Update functions that pass a FILE*.